+++ core/src/main/kotlin/org/pastalab/fray/core/RuntimeDelegate.kt
@@ -14,7 +14,7 @@ import java.util.concurrent.locks.ReentrantReadWriteLock
 import java.util.concurrent.locks.StampedLock
 import org.pastalab.fray.core.concurrency.HelperThread
 import org.pastalab.fray.core.utils.Utils.verifyOrReport
-import org.pastalab.fray.runtime.SyncurityCondition
+import org.pastalab.fray.runtime.RangerCondition
 
 class RuntimeDelegate(val context: RunContext) : org.pastalab.fray.runtime.Delegate() {
 
@@ -1133,13 +1133,13 @@ class RuntimeDelegate(val context: RunContext) : org.pastalab.fray.runtime.Deleg
     return 0
   }
 
-  override fun onSyncurityCondition(condition: SyncurityCondition) {
+  override fun onRangerCondition(condition: RangerCondition) {
     if (checkEntered()) {
       verifyOrReport(false, "This method should never be called recursively")
       return
     }
     try {
-      context.syncurityCondition(condition)
+      context.rangerCondition(condition)
     } finally {
       entered.set(false)
     }