+++ core/src/main/kotlin/org/pastalab/fray/core/ranger/AbortEvaluationException.kt
@@ -1,4 +1,4 @@
-package org.pastalab.fray.core.syncurity
+package org.pastalab.fray.core.ranger
 
 class AbortEvaluationException : RuntimeException {
   constructor() : super()